Namespace the appointments list table (drop the last global FFC_* class)
Namespace-compliance audit found one class out of line: the appointments
admin WP_List_Table subclass was defined in the GLOBAL namespace as
FFC_Appointments_List_Table, embedded inside
views/appointments-list.php — the only plugin class (besides the bootstrap
FFC_Autoloader, which must stay global) not under FreeFormCertificate\.

Moved it to its own autoloaded file
includes/self-scheduling/class-ffc-self-scheduling-appointments-list-table.php
as FreeFormCertificate\SelfScheduling\AppointmentsListTable (the autoloader's
SelfScheduling filename pattern resolves it). The class is self-contained;
the view now instantiates the namespaced class. The new file guard-requires
wp-admin/includes/class-wp-list-table.php so the \WP_List_Table parent is
available when the subclass autoloads.

Audit also confirmed: no global functions in includes/, and zero
namespace↔directory mismatches across all 26 sub-namespaces.

phpcs / PHPStan L8 clean; autoloader resolution verified; 389
self-scheduling/appointment tests green.
